+++ src/components/Movies/SavedMovies/SavedMovies.js
@@ -3,12 +3,10 @@ import SearchForm from "../SharchForm/SharchForm";
 import MoviesCardList from "../MoviesCardList/MoviesCardList";
 import Preloader from '../Preloader/Preloader';
 import api from "../../../utils/MainApi";
-import { useLocation } from 'react-router-dom';
 
 const SavedMovies = ({ movies, setMovies, savedMovies, setSavedMovies, searchMovie, searchError, setSearchError, checkSavedCards, setCheckSavedCards}) => {
   const [isSearched, setIsSearched] = useState(false);
-/*   const [search, setIsSearch] = useState('' || localStorage.getItem('savedSearchQuery')); */
-  const location = useLocation();
+  const [inputValue, setInputValue] = useState('' || localStorage.getItem('savedSearchText'));
   
     useEffect(() => {
       api.getSavedMovies(localStorage.getItem('jwt'))
@@ -29,10 +27,27 @@ const SavedMovies = ({ movies, setMovies, savedMovies, setSavedMovies, searchMov
     }
   }, [setCheckSavedCards]);
 
+  const handleSearch = (e) => {
+    setInputValue(e.target.value);
+  };
+
+  const handleSubmit = (evt) => {
+    evt.preventDefault();
+    setIsSearched(true)
+    setTimeout(() => setIsSearched(false), 1000);
+    setTimeout(() => searchMovie(inputValue), 1001);
+    setInputValue('');
+    localStorage.setItem('savedSearchText', inputValue);
+  }
+
   return (
     <section className="movies">
-      <SearchForm searchMovie={searchMovie} setIsSearched={setIsSearched} checkSavedCards={checkSavedCards} setCheckSavedCards={setCheckSavedCards}
-       /* onChange={handleSearch} search={search} onSubmit={handleSubmit} *//>
+      <SearchForm searchMovie={searchMovie}
+        setIsSearched={setIsSearched}
+        checkSavedCards={checkSavedCards}
+        setCheckSavedCards={setCheckSavedCards}
+        onChange={handleSearch}
+        onSubmit={handleSubmit}/>
         <h2 className={searchError ? 'movies__search-error' : 'movies__search-error movies__search-error-hidden'}>
           Ничего не найдено!
         </h2>
