+++ src/components/Movies/SharchForm/SharchForm.js
@@ -1,15 +1,9 @@
-import React, {useState} from "react";
+import React from "react";
 import FilterCheckbox from "../../FilterCheckbox/FilterCheckbox";
 import { useLocation } from 'react-router-dom';
 
 const SearchForm = ({ searchMovie, setIsSearched, checkSavedCards, setCheckSavedCards, checkbox, setCheckbox, onChange, onSubmit }) => {
-  const [inputValue, setInputValue] = useState('');
   const location = useLocation();
-
-/*   const handleCheckbox = () => {
-    localStorage.setItem("checkboxData", JSON.stringify(!checkbox));
-    setCheckbox(!checkbox);
-  }; */
 
   const handleCheckbox = () => {
     if (location.pathname === '/movies') {
@@ -22,26 +16,17 @@ const SearchForm = ({ searchMovie, setIsSearched, checkSavedCards, setCheckSaved
     }
   };
 
-   const handleChangeInput = (evt) => {
-    setInputValue(evt.target.value);
-    console.log(evt.target.value);
-}
-
-  const handleSubmit = (evt) => {
-    evt.preventDefault();
-    setIsSearched(true)
-    setTimeout(() => setIsSearched(false), 1000);
-    setTimeout(() => searchMovie(inputValue), 1001);
-    setInputValue('');
-  }
+  const setSearchValue = () => {
+    return location.pathname === '/movies' ? localStorage.getItem('searchText') : localStorage.getItem('savedSearchText');
+  };
 
   return (
     <form className="sharch-form">
       <fieldset className="sharch-form__content">
         <div className="sharch-form__box">
           <input className="sharch-form__input" type="search" placeholder="Фильм" id="film" required="false"
-            onChange={handleChangeInput} defaultValue = {/* setSearchValue() || */ ''} ></input>
-          <button className="sharch-form__button" onClick={handleSubmit}></button>
+            onChange={onChange} defaultValue = { setSearchValue() || ''} ></input>
+          <button className="sharch-form__button" onClick={onSubmit}></button>
         </div>
         <FilterCheckbox onClick={handleCheckbox} defaultChecked={checkbox} checkSavedCards={checkSavedCards} setCheckSavedCards={setCheckSavedCards}/>
       </fieldset>
